% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresse_henter.R
\name{adresse_henter}
\alias{adresse_henter}
\title{Title}
\usage{
adresse_henter(
  x,
  city_column = "address.city",
  country_column = "address.country"
)
}
\arguments{
\item{x}{Datasett, enten pull_x dataene eller All_Data}

\item{city_column}{kollone som inneholder bynavn}

\item{country_column}{kollone som inneholder landnavn}
}
\value{
En tibble med to kolloner, City og Country
}
\description{
Title
}
\examples{
\dontrun{
adresse_henter(Pull_2)
}
}
