% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruppe_summerer.R
\name{gruppe_summerer}
\alias{gruppe_summerer}
\title{gruppe_summerer}
\usage{
gruppe_summerer(name, fagGruppe, ...)
}
\arguments{
\item{name}{gir navn til ny kollone med summen av fagene innenfor denne}

\item{fagGruppe}{hvilket institutt som skal hentes}

\item{...}{kan velge om det ønskes publikasjoner eller siteringer, returneres til kattegori_henter funksjonen}
}
\value{
En data.table hvor fagene er redusert ned til kattegoriene bestemt i fagkattegorier filen
}
\description{
gruppe_summerer
}
\examples{
gruppe_summerer(All_Data)
}
