% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kattegori_henter.R
\name{kattegori_henter}
\alias{kattegori_henter}
\title{Kattegori henter}
\usage{
kattegori_henter(Institutt, type = "publications")
}
\arguments{
\item{Institutt}{hvilket institutt eller fakultet som skal hentes ut}

\item{type}{publikasjoner eller siteringer}
}
\value{
data.frame med fagene som tilhører et institutt eller faktultet
}
\description{
Kattegori henter
}
\examples{
\dontrun{
kattegori_henter("political science", type = "publications")
}
}
